Move test functions out of lib
